MUST_REVERT Do not alter the wpa_supplicant.conf file
The libhardware_legacy support for managing the
wpa_supplicant.conf file was not only hard-coded to use the
"wifi.interface" property as a non-directory but also to
over-write the wpa_supplicant.conf file to that value
regardless of whether that was the right thing to do.  This
change simply comments out the range of code to get things
working in a hurry.

In future, we need to figure out the right fix for this that
preserves the over-writing behavior in the (unknown?) case
where that was desirable, while preventing it when not
desired.

Change-Id: I77bef1096a6b0df798c1a4325f0a97d001f6dc1b
Signed-off-by: Paul Drews <paul.drews@intel.com>