WIFI: load and unload modules by dependency
If the WIFI_DRIVER_PROBE_PATH is defined with a non-empty string,
WIFI HAL layer will call insmod_by_dep() in libcutils to load wifi
driver module and other modules it depends on, and try to remove
all of these modules from kernel by calling rmmod_by_dep() when it
is asked at run time.

The old approach with a single insmod() is still kept as a backup
when probe path is not provided.

This commit also includes other several changes to refine the
code around module operations.

Change-Id: I48db96890bf5ebc2ec2ec3896b50585b87ceb29a
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>