Restore the ability to use ctrl_interface directory
lihardware_legacy was ALWAYS re-writing the
wpa_supplicant.conf file's "ctrl_interface" entry to the
value specified as the property wifi.interface, whose value
was typically "wlan0".  This was designed to allow the
wpa_supplicant service to be started with a "socket"
directive in its service description in an init.rc file.
Under these circumstances, the socket name is exchanged
between wpa_supplicant and its WifiService client through an
environment variable whose name is derived from
wifi.interface and a standard prefix.

However, since the "ctrl_interface" entry was always
re-written in this way, it broke the ability to use
"ctrl_interface" in its standard usage as a directory where
the (unix-domain) socket will be exchanged.

This change restores the ability to use the standard
directory-form ctrl_interface usage by limiting the
re-writing of the ctrl_interface entry to cases where it
does not look like an absolute path or the "DIR=" form of
directory.

Issue: AXIA-436
Change-Id: I2d51bd49a501cbcd38def8e3984b544dbca54d98
Signed-off-by: Paul Drews <paul.drews@intel.com>