Revert "MUST_REVERT wait for sysfs nodes on driver-load"
This reverts commit
66f3d335990f0222aa6e323ab8bb089916873584.

The patch being reverted was a quick-and-dirty change to
wait for the sysfs nodes exported by the wifi driver to be
created before reporting completion of driver-load.  These
sysfs nodes are needed by the client of the driver.

The patch is being reverted in preparation for a replacement
patch that does the same thing but in a more robust way.

For: AXIA-437
Change-Id: I7072165cdb24d940cc04335de97313dc5ef7760e
Signed-off-by: Paul Drews <paul.drews@intel.com>