wifi: Add ability to use custom wifi code
Adds a switch so that a different source file for wifi can be
used instead of the default one from AOSP. This gives us the
flexibility of having different wifi code for various build
targets.

One may argue that the code can be built as a static library
and be included here. However, since nothing in here refers
to functions in the wifi code, the linker simply does not
include the library. So the source has to be built at the
same time as other source files here.

Also by moving the code away from this project, this makes
it easier for us to move forward to upstream commits.

For: AXIA-1242
Change-Id: Icf3851590742788c23eca12ca3344515ea6f1fcf
Signed-off-by: Daniel Leung <daniel.leung@intel.com>